// 04_desafios_frontend/2_challenge/src/components/types.ts
export type Temperature = "Celsius (°C)" | "Fahrenheit (°F)";
export type WindSpeed = "km/h" | "mph";
export type Precipitations = "Millimeters (mm)" | "Inches (in)";
export type UnitOfMesure = "Imperial" | "Metric";
export type MetricObject = {
  sistem: UnitOfMesure;
  temperature: Temperature;
  windPeed: WindSpeed;
  precipitation: Precipitations;
};